requirements.txt
flask
flask-CORS
SpeechRecognition
googletrans
pydub
dotenv
gtts
google-generativeai
